add script that for downloading compute node packages
For compute nodes that do not have access to the internet in HPC systems, the
head node, that has internet access, should download the packages that the
compute nodes require and give it to them to install offline.

The head node uses `apt-rdepends` to find all the required package's
dependencies and then downloads the .deb files for the package and all its
dependencies. The .deb files are then stored in a specific folder that will be
copied/moved to the network file system for the compute nodes to access.
This script automates that process.